requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiokafka>=0.11.0
python-dotenv>=1.0.1
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0

langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langgraph>=0.2.0
pinecone>=5.0.0
